Fix the triggering of the github workflow happening on branches
Prior to this change, the workflow intended to prepare new releases from tags were being executed every time a branch was created.

This change sets the new release workflow to be triggered on `push` events ignoring all branch references. As it turns out, it is not possible at the moment to filter the origin of `create` events.

References:
- https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#onpushpull_requestbranchestags