Change push authentication on the tag release workflow
Prior to this change, the push was being authenticated with the default `GITHUB_TOKEN` generated for every action.
This is not supported as the default token does not have permission to push to protected branches, even if the workflow is triggered from the same branch.

This change alters the action to use a secret `BOT_TOKEN` to authenticate the git push. This secret is a Private Access Token of a bot account with the required permissions

The issue is discussed in this thread on the forum:
- https://github.community/t5/GitHub-Actions/How-to-push-to-protected-branches-in-a-GitHub-Action/td-p/29609