Fix constant declaration to prevent SSR errors
The limit is currently defined in compile time by replacing a token string with the actual value. To avoid runtime errors, the limit is specified as a string and converted at runtime to a number. This causes an error when pre-rendering SSR applications like Next.js because `window` does not exist.